This should actually be 32 bit. Maybe even 16.

+++ stop_watch.go
@@ -8,6 +8,6 @@ type StopWatch struct {
 	start, stop time.Time
 }
 
-func (self *StopWatch) Milliseconds() uint64 {
-	return uint64(self.stop.Sub(self.start) / time.Millisecond)
+func (self *StopWatch) Milliseconds() uint32 {
+	return uint32(self.stop.Sub(self.start) / time.Millisecond)
 }
